Saving months in databse and displaying all dance info on homepage
- Combined street address/location into one input because there is
  only one database field to hold the info

...Days of the week not currently stored in database, looks like month
not being stored either, it's just being set to '0'.

...The selector for sending months in the Ajax request doesn't
actually exist, so I'm accumulating all of the selected months in a
variable and I'll make that a string to submit to the database.

Ok, each month is a key in a dictionary with the position as the
value. ex {'January':0}. So I get the changed checkbox, find the index
of the month in the list of months, and toggle either the month name
or an empty string in the list. This way all of the months are kept in
order, and then I have a little loop to combine all of the strings
that are not ''. This is the "months" value that I'll send to the
database.

...months are saving and displaying