Edit button links to edit entry route
`dance/<entry_id>/edit` edits dance entry by ID.

---

Got rid of all of the JS stuff I was doing to edit existing
entries... I was going to re-use the modal and populate it with
values, but now I'm just going to use a normal form, so I can just
fill the values in with template variables. I told Dimitriy that this
way is better, because he'll have more control when editing entries,
ex. the types of dances are just stored as a text field. In the modal
they are a bunch of checkboxes, but when editing, it will just be a
text field with the actual text that is stored in the database, and if
the admin wants to add a type of dance that is not provided by the
modal's checkboxes, it can just be typed in.